+++ README.md
@@ -3,6 +3,7 @@
 ## Описание:
 Этот проект рабоатет с банковским операциями,
 позволяя сделать маску карты или счета.
+генераторы-позволяют узнать id операции и тип операции
 
 ## Установка:
 1.Клонируйте репозиторий:
@@ -11,7 +12,7 @@
 2. Используйте локальный сервер
 
 ## Тесты:
-покрытие тестов 86%
+покрытие тестов 90%
 
 ### проверка:
 poetry run pytest --cov  — через poetry.
