+++ main.py
@@ -1,25 +1,6 @@
-import pytest
+from src.generators import filter_by_currency, transaction_descriptions, card_number_generator
 
-
-@pytest.fixture
-def state():
-    return [
-        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-        {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
-    ]
-
-
-@pytest.fixture
-def data_state():
-    return [
-        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-        {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
-        {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
-        {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
-    ]
-@pytest.fixture
-def user_tansaction():
-    return [{"id": 939719570,
+transactions = [{"id": 939719570,
                  "state": "EXECUTED",
                  "date": "2018-06-30T02:08:58.425572",
                  "operationAmount":
@@ -78,3 +59,13 @@ def user_tansaction():
                     "to": "Счет 14211924144426031657",
                 },
                 ]
+
+for trans in filter_by_currency(transactions, "USD"):
+    print(trans)
+
+descriptions = transaction_descriptions(transactions)
+for key in range(5):
+    print(next(descriptions))
+
+for card_number in card_number_generator(1, 5):
+    print(card_number)
