--- tests/generators.py
@@ -1,31 +1,0 @@
-from src.generators import card_number_generator, transaction_descriptions
-
-
-def test_transaction_descriptions():
-    transactions = [
-        {"description": "Перевод организации"},
-        {"description": "Перевод со счета на счет"},
-        {"description": "Перевод со счета на счет"},
-        {"description": "Перевод с карты на карту"},
-        {"description": "Перевод организации"},
-    ]
-
-    descriptions = list(transaction_descriptions(transactions))
-    expected = [
-        "Перевод организации",
-        "Перевод со счета на счет",
-        "Перевод со счета на счет",
-        "Перевод с карты на карту",
-        "Перевод организации",
-    ]
-
-    assert descriptions == expected
-
-
-def test_card_number_generator():
-    generator = card_number_generator(1, 5)
-    assert next(generator) == "0000 0000 0000 0001"
-    assert next(generator) == "0000 0000 0000 0002"
-    assert next(generator) == "0000 0000 0000 0003"
-    assert next(generator) == "0000 0000 0000 0004"
-    assert next(generator) == "0000 0000 0000 0005"
